refactor(ci): split the latency gate so no file carries 800 lines
gate-rule-latency.ts had grown to 867 lines and its test to 1024, both past the
project's 800-line ceiling, and most of that weight was justification rather
than machinery.

Thresholds move to scripts/lib/rule-latency-thresholds.ts, which is nothing but
six constants and the measurements they came from, so a pull request that
loosens one reads as a diff to a file whose entire contents are the argument
against doing that, rather than three characters inside 800 lines of plumbing.

Tests split by what they touch: gate-rule-latency.test.ts stays synthetic, while
gate-rule-latency-guards.test.ts holds the ones that drive a real engine, a real
git repository, the committed baseline and the CLI process boundary. Shared
profile builders move to tests/rule-latency-fixtures.ts rather than being copied.

No behaviour change; 104 tests pass. MEASUREMENT_PATHS, CODEOWNERS and the
workflow path filters are widened to the new files, so that editing a threshold
still puts the whole corpus back in scope and still needs TSC review.
