+++ .github/CODEOWNERS
@@ -52,8 +52,9 @@ scripts/check-rules-safety.ts  @Agent-Threat-Rule/tsc @eeee2345
 # for the same reason: shrinking the measurement basis mutes the latency
 # gate without touching either the gate or its baseline.
 scripts/gate-rule-latency.ts          @Agent-Threat-Rule/tsc @eeee2345
-scripts/lib/rule-latency-profile.ts   @Agent-Threat-Rule/tsc @eeee2345
-tests/gate-rule-latency.test.ts       @Agent-Threat-Rule/tsc @eeee2345
+scripts/lib/rule-latency-*.ts         @Agent-Threat-Rule/tsc @eeee2345
+tests/gate-rule-latency*.test.ts      @Agent-Threat-Rule/tsc @eeee2345
+tests/rule-latency-fixtures.ts        @Agent-Threat-Rule/tsc @eeee2345
 scripts/gate-re2-portability.ts       @Agent-Threat-Rule/tsc @eeee2345
 scripts/eval-generalization.ts        @Agent-Threat-Rule/tsc @eeee2345
 data/rule-latency-baseline.json       @Agent-Threat-Rule/tsc @eeee2345
